Add the ability to add buttons
Advanced toasts (requiring Windows 10, I believe) allow adding buttons:
https://docs.microsoft.com/en-us/windows/uwp/design/shell/tiles-and-notifications/adaptive-interactive-toasts

Let's do the same, via a new helper: addAction(label). This also
requires an update to the IWinToastHandler interface.

Note that we need to change the type to "ToastGeneric", as per the
documentation: actions would work when keeping the Template* type for
now, but nothing guarantees that this will keep working.

As we now change the type, we also have to set the duration
attribute specifically (otherwise it seems to default to "short",
which is about 7 seconds).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>