Optionally set an expiration time
Obviously, the application could set a timer and then hide the
notification. That is not quite the same as asking Windows properly to
please expire the toast after a given time.

Sadly, this is a bit more complicated than one would expect because
there is no convenience IReference<T> wrapper which we could use to wrap
a DateTime object in the form put_ExpirationTime() expects.

Another bummer: the dismissal reason is unexpectedly UserCanceled if
the toast expires, not TimedOut. Let's work around that in this library
and detect whether the expiration timeout really was reached, and
pretend that the dismissal reason was TimedOut in that case.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>