Refactor out the function to create the shortcut explicitly
In the experiments of this developer, it would appear that toast
notifications are not shown reliably if the shortcut has just been
created. This behavior was reproduced with console and windows
applications, with WinToast and with other toast notification examples.

The only workaround that seemed to show the notifications after a
shortcut was created: sleep for five seconds.

Let's give callers a chance to create the shortcut explicitly, so that
this can be performed, say, in the install phase of an application, so
that the shortcut exists for a lot longer than just five seconds once
the toast notification needs to be shown to the user.

To give the caller even more control, we declare a new enum type that
describes exactly what happened, with several error conditions (negative
values) and with a couple of (non-negative) values describing what
exactly happened.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>