Console Example: add an ugly workaround for the first call
The first call necessarily adds a new shortcut, and the toast may then
not be shown correctly. At least on this developer's machine, it is not
shown at all just after the shortcut was created.

However, it works if we first create the shortcut, wait for 3s to give
Windows a chance to parse the shortcut, and then just spawn the same
executable, with the same options, again.

This is far from elegant, but at least it works...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>