Console Example: add an option to only create the shortcut
The first call necessarily adds a new shortcut, and the toast may then
not be shown correctly. At least on this developer's machine, it is not
shown at all just after the shortcut was created, unless an ugly
Sleep(5000) is inserted (to wait five seconds).

So let's add an option to the command line example where the shortcut
corresponding to App name and the AppModelUserID is created, and nothing
else.

That way, scripts that use this example can first ensure that the
shortcut exists (e.g. in some installation phase of the software), and
show the toast notification only much later.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>