Fix the incorrect test in the example whether showToast() failed
In our Qt example, we test specifically whether showToast() failed by
looking at the return value. However, we tested whether the return value
is 0, when already the first error condition (when WinToast was not
initialized) actually returns -1.

It is also safer to use a negative value as error condition because id
could conceivably be set to 0 legitimately (00000000 is a valid first
part of a GUID).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>