vcxproj: reorder settings into simpler sections
This commit separates the settings into five sections: settings that
apply to all configurations and platforms, settings that apply to the
Win32 platform, settings that apply to the x64 platform, settings that
apply to the Debug configuration and settings that apply to the Release
configuration.

Amongst others, this fixes the issue where a 64-bit release build still
spat out debug messages.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>